refactor: ステージ進捗の判定を分離
